refactor(ansible): rewrite include_tasks module
This commit refactors the Ansible tasks and variables for Terraform installation.
The ansible.builtin.include_tasks task is replaced with a more flexible lookup
function that searches for the first available file in a list of possible files.
The base_os variable is removed, and the params variable is introduced to specify
the files and paths to search for. This makes the installation process more robust
and easier to maintain.